refactor(daemon): replace Image with DockerImage
